pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fitness-agent"
version = "0.1.0"
description = "Speediance AI 私教智能体 — 基于 LangGraph 的多意图健身助手"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Speediance AI Team" },
]

dependencies = [
    # ── LangChain / LangGraph ──────────────────────────────────
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",          # Qwen 走 OpenAI 兼容接口
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langgraph-checkpoint-redis>=0.0.1",   # Redis 短期记忆 checkpointer

    # ── Web 框架 ───────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # ── 数据库 ────────────────────────────────────────────────
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "psycopg[binary,pool]>=3.2.0",      # langgraph-checkpoint-postgres 依赖

    # ── 缓存 ─────────────────────────────────────────────────
    "redis[asyncio]>=5.0.0",
    "hiredis>=2.3.0",                   # redis 性能加速

    # ── 配置管理 ──────────────────────────────────────────────
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # ── HTTP 客户端 ───────────────────────────────────────────
    "httpx>=0.27.0",
    "tenacity>=8.3.0",                  # 重试装饰器

    # ── 工具库 ────────────────────────────────────────────────
    "python-dateutil>=2.9.0",           # 时间解析
    "pytz>=2024.1",
    "structlog>=24.1.0",                # 结构化日志
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",                    # FastAPI TestClient
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
fitness-agent = "fitness_agent.main:cli_main"

# ── hatch 环境 ────────────────────────────────────────────────
[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
dev     = "uvicorn fitness_agent.main:app --reload --host 0.0.0.0 --port 8000"
test    = "pytest tests/ -v --asyncio-mode=auto"
lint    = "ruff check src/ tests/"
format  = "ruff format src/ tests/"
typecheck = "mypy src/"

[tool.hatch.build.targets.wheel]
packages = ["src/fitness_agent"]

# ── ruff ──────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

# ── mypy ─────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

# ── pytest ────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
